Add scrapers, bot, SQL schema setup script
The scraper is configured to use the popular
nekretnine.rs website. It takes the configuration
to run from ini files in its directory.
It then outputs the interesting real estate to a
SQLite database file.

The SQL engine is chosen as SQLite as this is
a personal project which has no significant traffic.
It is more important to have this portable than highly
available, hence SQLite.

The bot is chosen to be a Telegram bot for its ease of
use and my usage of Telegram. It reads the configuration
from appsettings*.json files to avoid leaking credentials
such as the bot token and to provide a configurable place
for storing the database.

The gitignore file was mostly copy-pasted from github/gitignore,
with some manual additons to the top of the file to enable local
configurations of the scripts which are to be run.