TelegramBot: Move to Azure Key Vault
The configuration files which are keeping secrets are difficult
to manage via automated deployments. Since this application is
to be deployed and run on a cloud platform, it is easier to manage
keys remotely. The internet access is required either way since this
application is sending messages via Telegram API.

Azure was chosen as the go-to cloud provider for this application,
so Azure Key Vault makes most sense to use.

Closes #3.