DB: Switch from SQLite to Cosmos DB
Since the project is to be hosted on Azure, the database
will benefit from the availability/scalability of an Azure
service. The current setup with SQLite works but is not
persistent over container restarts as the disk is restarted
as well. Cosmos DB ensures the Telegram notifier can send
only what is needed as it is persisted.